// frontend/src/types/Insight.ts
export interface Insight {
  _id: string;
  end_year?: number;
  intensity?: number;
  sector: string;
  topic: string;
  insight: string;
  url: string;
  region: string;
  start_year?: number;
  impact?: number;
  added: string;
  published: string;
  country: string;
  relevance?: number;
  pestle: string;
  source: string;
  title: string;
  likelihood?: number;
  __v: number;
}